More rigorous correctness check
